pyproject.toml
[tool.poetry]
name = "selfbill"
version = "0.1.0"
description = "Empty project"
authors = ["mamashin <mamashin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.0"
loguru = "^0.7.2"
python-decouple = "^3.5"
djangorestframework = "^3.13.1"
httpx = "^0.26.0"
sentry-sdk = "^1.5.2"
redis = "^5.0.1"
rq = "^1.10.1"
django-rq = "^2.5.1"
pillow = "^10.1.0"
django-behaviors = "^0.5.1"
django-split-settings = "^1.2.0"
django-ipware = "^6.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
